use parameters file for process arguments
This is supported from Java 9 onwards
